Removes the dead SQLAlchemy metric value query path
The backend now serves metric values through the climate_ref.results.Reader
facade. The old hand-rolled query path in core/metric_values.py and
core/outliers.py was never deleted after that migration and had no
remaining callers.

Moves MetricValueType and parse_id_list, the only symbols from
metric_values.py still in use, into core/reader_values.py and removes
the now-empty module along with outliers.py and the unused
AnnotatedScalarValue model.
