+++ backend/src/ref_backend/api/routes/diagnostics.py
@@ -10,13 +10,11 @@ from climate_ref.models.dataset import CMIP6Dataset
 from climate_ref.results import MetricValueFilter
 from ref_backend.api.deps import AppContextDep
 from ref_backend.core.filter_utils import build_filter_clause
-from ref_backend.core.metric_values import (
-    MetricValueType,
-    parse_id_list,
-)
 from ref_backend.core.reader_values import (
+    MetricValueType,
     fetch_metric_values,
     parse_dimension_filters,
+    parse_id_list,
 )
 from ref_backend.models import (
     Collection,
