+++ changelog/48.trivial.md
@@ -1,0 +1,1 @@
+Removed the old hand-rolled SQLAlchemy query path for metric values, now that the API is served by the `climate_ref.results.Reader` facade.
